+++ Changelog.md
@@ -44,6 +44,8 @@ Deprecations:
   `RSpec::Core::Configuration#warnings?`. (Myron Marston)
 * Deprecate `share_examples_for` in favor of `shared_examples_for` or
   just `shared_examples`. (Myron Marston)
+* Deprecate `RSpec::Core::CommandLine` in favor of
+  `RSpec::Core::Runner`. (Myron Marston)
 
 ### 2.99.0.beta2 / 2014-02-17
 [full changelog](http://github.com/rspec/rspec-core/compare/v2.99.0.beta1...v2.99.0.beta2)